requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0
sqlparse==0.5.0
setuptools==75.2.0
django-extensions==3.2.3
Pillow==11.0.0
requests==2.32.3